+++ modules/core/quantity/quantity.post_update.php
@@ -6,6 +6,7 @@
  */
 
 use Drupal\Core\Entity\Entity\EntityViewMode;
+use Drupal\system\Entity\Action;
 
 /**
  * Create plain text view mode for quantities.
@@ -28,4 +29,23 @@ function quantity_post_update_plain_text_view_mode(&$sandbox) {
     ],
   ]);
   $view_mode->save();
+}
+
+/**
+ * Create quantity delete action.
+ */
+function quantity_post_update_delete_action(&$sandbox) {
+  $action = Action::create([
+    'id' => 'quantity_delete_action',
+    'label' => t('Delete quantity'),
+    'type' => 'quantity',
+    'plugin' => 'entity:delete_action:quantity',
+    'configuration' => [],
+    'dependencies' => [
+      'module' => [
+        'quantity',
+      ],
+    ],
+  ]);
+  $action->save();
 }